feat: 감사 5건 전수 대응 — drift 롤백 + EVM trace + 블록 서브배치 + BTC batch + 주소 백필
- [Critical] drift 롤백: block-scan 모드에서 chain-wide rollback (fetchReorgRollbackEvents 재사용)
- [Critical] EVM trace: debug_traceBlockByNumber callTracer 통합, 3단계 발견 전략 (direct → log → trace)
- [Medium] 블록 서브배치: ScanBlocks 50블록 단위 RPC 호출 (메모리 피크 ~75%↓)
- [Medium] BTC batch: 기본값 3→6 (prefetchPrevouts 최적화 후 안전)
- [Critical] 주소 백필: migration 025, backfillCh 파이프라인 패턴, Admin API POST /admin/v1/backfill

+++ internal/store/postgres/integration_test.go
@@ -88,7 +88,7 @@ func TestTokenRepo_UpsertAndFind(t *testing.T) {
 	require.NotNil(t, found)
 	assert.Equal(t, tokenID, found.ID)
 	assert.Equal(t, "TST", found.Symbol)
-	assert.Equal(t, int32(9), found.Decimals)
+	assert.Equal(t, 9, found.Decimals)
 
 	// Idempotent upsert returns same ID.
 	tx2, err := db.BeginTx(ctx, nil)
@@ -665,9 +665,9 @@ func TestBalanceEventRepo_RollbackDeleteByCursor(t *testing.T) {
 			ActivityType: model.ActivityDeposit, EventAction: "deposit",
 			Address: watchedAddr, Delta: "100000",
 			WatchedAddress: &watchedAddr, BlockCursor: cursor,
-			ChainData:      json.RawMessage("{}"),
-			EventID:        fmt.Sprintf("rollback-evt-%d-%s", cursor, uuid.NewString()[:8]),
-			FinalityState:  "confirmed", BalanceBefore: strPtr("0"),
+			ChainData:     json.RawMessage("{}"),
+			EventID:       fmt.Sprintf("rollback-evt-%d-%s", cursor, uuid.NewString()[:8]),
+			FinalityState: "confirmed", BalanceBefore: strPtr("0"),
 			BalanceAfter: strPtr("100000"), DecoderVersion: "v1", SchemaVersion: "v1",
 		})
 		require.NoError(t, uErr)
